Rework gravity to use G's
Before now, we decided to apply gravity based on a global gametic.
Now, we add parts of a G over several tics, and every time we reach our
G value we move the block down one or more spaces.

This commit also cleaned up the main frame function to save a table
containing player state to a local for reuse throughout the function.