Rework serialization to deal with references
Our old serialization was naive and did not account for tables or
userdata with multiple references.  The new strategy is to first
create a flat map that contains references for keys and an id as a
value.  Then, we serialize that flat map except we reverse things - the
id is now the key and table or userdata contents is the value.  And
when we serialize a table, we serialize any references to other tables
or userdata in that table as a simple little array containing an id.